requirements.txt
# it is recommended to specify library versions
pyyaml
torch==2.0.1
torchvision==0.15.2
numpy==1.25.2
scikit-learn==1.3.1
Pillow==10.0.0
pandas==2.1.0
opencv-python>=4.6.0
ultralytics
openmim
joblib

--index-url https://download.pytorch.org/whl/cpu --extra-index-url https://pypi.org/simple